+++ packages/eyes-storybook/test/fixtures/appWithStorybook/index.js
@@ -97,6 +97,7 @@ storiesOf('skipped tests', module)
   )
   .add('[SKIP] this story should not be checked visually by eyes-storybook because of global config',
     () => <div>this story should not be checked visually by eyes-storybook because of global config</div>)
+  .add('testing circular parameters', () => <div>nothing to see here</div>, {circular, eyes: {include: false}})
 
 storiesOf('Text', module)
   .add(