+++ packages/eyes-storybook/test/it/getStories.it.test.js
@@ -28,6 +28,10 @@ describe('getStories', () => {
     try {
       await page.goto('http://localhost:9001');
       const stories = await page.evaluate(getStories);
+      const parameters = {
+        fileName: './test/fixtures/appWithStorybook/index.js',
+        framework: 'react',
+      };
       expect(stories).to.eql(
         [
           {
@@ -39,6 +43,7 @@ describe('getStories', () => {
             name: 'with text',
             kind: 'Button',
             parameters: {
+              ...parameters,
               someParam: 'i was here, goodbye',
               eyes: {ignoreRegions: [{selector: '.ignore-this'}]},
             },
@@ -48,39 +53,56 @@ describe('getStories', () => {
             kind: 'Button',
             error: `Ignoring parameters for story: "with some emoji Button" since they are not serilizable. Error: "Converting circular structure to JSON\n    --> starting at object with constructor 'Object'\n    --- property 'inner' closes the circle"`,
           },
-          {name: 'image', kind: 'Image'},
-          {name: 'story 1', kind: 'Nested'},
-          {name: 'story 1.1', kind: 'Nested/Component'},
-          {name: 'story 1.2', kind: 'Nested/Component'},
-          {name: 'a yes-a b', kind: 'Button with-space yes-indeed'},
-          {name: 'b yes-a b', kind: 'Button with-space yes-indeed/nested with-space yes'},
+          {name: 'image', kind: 'Image', parameters},
+          {name: 'story 1', kind: 'Nested', parameters},
+          {name: 'story 1.1', kind: 'Nested/Component', parameters},
+          {name: 'story 1.2', kind: 'Nested/Component', parameters},
+          {name: 'a yes-a b', kind: 'Button with-space yes-indeed', parameters},
+          {
+            name: 'b yes-a b',
+            kind: 'Button with-space yes-indeed/nested with-space yes',
+            parameters,
+          },
           {
             name: 'c yes-a b',
             kind: 'Button with-space yes-indeed/nested with-space yes/nested again-yes a',
+            parameters,
           },
-          {name: 'story 1.1', kind: 'SOME section|Nested/Component'},
-          {name: 'story 1.2', kind: 'SOME section|Nested/Component'},
+          {name: 'story 1.1', kind: 'SOME section|Nested/Component', parameters},
+          {name: 'story 1.2', kind: 'SOME section|Nested/Component', parameters},
           {
             name: 'c yes-a b',
             kind: 'Wow|one with-space yes-indeed/nested with-space yes/nested again-yes a',
+            parameters,
           },
-          {name: 'should also do RTL', kind: 'RTL'},
-          {name: 'local RTL config', kind: 'RTL', parameters: {eyes: {variations: ['rtl']}}},
+          {name: 'should also do RTL', kind: 'RTL', parameters},
+          {
+            name: 'local RTL config',
+            kind: 'RTL',
+            parameters: {...parameters, eyes: {variations: ['rtl']}},
+          },
           {
             name:
               'this story should not be checked visually by eyes-storybook because of local parameter',
             kind: 'skipped tests',
-            parameters: {eyes: {include: false}},
+            parameters: {...parameters, eyes: {include: false}},
           },
           {
             name:
               '[SKIP] this story should not be checked visually by eyes-storybook because of global config',
             kind: 'skipped tests',
+            parameters,
+          },
+          {
+            name: 'testing circular parameters',
+            kind: 'skipped tests',
+            parameters: {eyes: {include: false}}, // note that fileName and framework parameters are not present here because of the circular reference
           },
           {
             kind: 'Text',
             name: 'appears after a delay',
             parameters: {
+              ...parameters,
               eyes: {
                 waitBeforeScreenshot: '.ready',
               },
@@ -89,31 +111,19 @@ describe('getStories', () => {
           {
             name: 'Popover',
             kind: 'Interaction',
-            parameters: {bgColor: 'lime', eyes: {runBefore: '__func'}},
+            parameters: {...parameters, bgColor: 'lime', eyes: {runBefore: '__func'}},
           },
           {
             kind: 'Responsive UI',
             name: 'Red/green',
+            parameters,
           },
         ].map((story, index) => {
-          const {name, kind, parameters, error} = story;
-          const res = {
+          return {
+            ...story,
             index,
-            name,
-            kind,
             isApi: true,
           };
-          if (!error) {
-            res.parameters = {
-              fileName: './test/fixtures/appWithStorybook/index.js',
-              framework: 'react',
-              ...parameters,
-            };
-          } else {
-            res.error = error;
-          }
-
-          return res;
         }),
       );
     } finally {